Handle SIGTTOU by disabling read/write
If the process using the fesvr (e.g. Spike) is run in the
background, then calling tcsetattr() causes SIGTTOU to be received
and further execution is blocked.

This commit adds a handler for SIGTTOU which, if called, turns the
canonical_termios_t::read and write functions into no-ops. This is
sufficient to allow execution to proceed if the process is already
backgrounded when fesvr is loaded (e.g. if launched on the
terminal with &, or launched by creduce).